Moved tuning submenu to its own file
